Nerf Shana's Halo, buff Specialize (Shana)
Shana's Halo is literally broken, so limit it to one, in return buff
Specialize to make up for the lack of "3" halos.

It's broken in the sense that the game only checks if you need all 3
pieces once, so this will end up in the player having 7 halos in the
end.